Refactor: Decompose app.py into smaller modules
I've refactored the main application file, app.py, into
a set of more focused modules under the src/filekitty directory.
This includes separate modules for constants, utilities, Qt widgets,
dialogs, Python parsing logic, history management, the main window,
and application startup logic.

This change aims to improve code organization, maintainability,
and readability as part of issue #13.
