+++ src/filekitty/__main__.py
@@ -1,0 +1,5 @@
+# src/filekitty/__main__.py
+from .app_logic import main
+
+if __name__ == "__main__":
+    main()
