+++ src/filekitty/constants.py
@@ -1,0 +1,8 @@
+ICON_PATH = "assets/icon/FileKitty-icon.png"
+HISTORY_DIR_NAME = "FileKittyHistory"
+STALE_CHECK_INTERVAL_MS = 2500  # Check every 2.5 seconds
+HASH_ERROR_SENTINEL = "HASH_ERROR"
+HASH_MISSING_SENTINEL = "FILE_MISSING"
+SETTINGS_DEFAULT_PATH_KEY = "defaultPath"
+SETTINGS_HISTORY_PATH_KEY = "historyPath"
+TEXT_CHECK_CHUNK_SIZE = 1024  # Bytes to read for text file check
